refactor: simplify usePanZoom hook by extracting handlers and events
