+++ src/hooks/usePanZoom.ts
@@ -9,12 +9,9 @@ import {
 import type { XAxisConfig, YAxisConfig } from "../services/persistence";
 import { getAxisById } from "../utils/axisCalculations";
 import { screenToWorld } from "../utils/coords";
-import {
-	applyZoomBoxToAxes,
-	applyZoomToRange,
-	computePinchGesture,
-	panRangeByPixels,
-} from "./panZoomMath";
+import { applyZoomToRange, panRangeByPixels } from "./panZoomMath";
+import { usePanZoomEvents } from "./usePanZoomEvents";
+import { usePanZoomHandlers } from "./usePanZoomHandlers";
 import { usePanZoomKeyboard } from "./usePanZoomKeyboard";
 
 interface UsePanZoomOptions {
@@ -356,103 +353,32 @@ export function usePanZoom({
 		],
 	);
 
-	const handleWheel = useCallback(
-		(e: React.WheelEvent, target: PanTarget = "all") => {
-			setIsWheeling(true);
-			panStateRef.current.active = true;
-			if (wheelTimeoutRef.current) clearTimeout(wheelTimeoutRef.current);
-			wheelTimeoutRef.current = setTimeout(() => {
-				setIsWheeling(false);
-				panStateRef.current.active = false;
-				onPanEnd();
-				wheelTimeoutRef.current = null;
-			}, 300);
-
-			const zoomFactor = e.deltaY > 0 ? 1.1 : 0.9;
-			smoothZoomRef.current = true;
-			const rect = containerRef.current?.getBoundingClientRect();
-			if (rect) containerRectRef.current = rect;
-			performZoom(
-				zoomFactor,
-				rect ? e.clientX - rect.left : width / 2,
-				rect ? e.clientY - rect.top : height / 2,
-				target,
-				e.shiftKey,
-			);
-		},
-		[containerRef, width, height, performZoom, onPanEnd, panStateRef, smoothZoomRef],
-	);
-
-	const handleMouseDown = useCallback(
-		(e: React.MouseEvent, target: PanTarget = "all") => {
-			const rect = containerRef.current?.getBoundingClientRect();
-			if (!rect) return;
-			containerRectRef.current = rect;
-			const x = e.clientX - rect.left,
-				y = e.clientY - rect.top;
-			if (e.ctrlKey && target === "all") {
-				if (
-					x >= padding.left &&
-					x <= width - padding.right &&
-					y >= padding.top &&
-					y <= height - padding.bottom
-				) {
-					const box = { startX: x, startY: y, endX: x, endY: y };
-					zoomBoxStartRef.current = box;
-					setIsZooming(true);
-					updateZoomBoxDom(box);
-				}
-			} else {
-				// A drag must track the cursor 1:1 — stop any running zoom easing.
-				smoothZoomRef.current = false;
-				setPanTarget(target);
-				panTargetRef.current = target;
-				shiftDownRef.current = e.shiftKey;
-				lastMousePos.current = { x: e.clientX, y: e.clientY };
-			}
-		},
-		[containerRef, padding, width, height, updateZoomBoxDom, smoothZoomRef],
-	);
-
-	const handleTouchStart = useCallback(
-		(e: React.TouchEvent, target: PanTarget = "all") => {
-			const now = Date.now(),
-				isDouble = now - lastTouchTime.current < 300;
-			lastTouchTime.current = now;
-			const rect = containerRef.current?.getBoundingClientRect();
-			if (rect) containerRectRef.current = rect;
-
-			if (e.touches.length === 1) {
-				const t = e.touches[0];
-				if (!rect) return;
-				if (isDouble) {
-					if (target === "all") {
-						handleAutoScaleX();
-						activeYAxes.forEach((a) => {
-							handleAutoScaleY(a.id);
-						});
-					} else if (typeof target === "object") {
-						if ("xAxisId" in target) handleAutoScaleX(target.xAxisId);
-						else if ("yAxisId" in target)
-							handleAutoScaleY(target.yAxisId, t.clientY - rect.top);
-					}
-					return;
-				}
-				setPanTarget(target);
-				lastTouchPos.current = { x: t.clientX, y: t.clientY };
-			} else if (e.touches.length === 2) {
-				setPanTarget((prev) => (prev && prev !== "all" ? prev : target));
-				const t1 = e.touches[0],
-					t2 = e.touches[1];
-				lastPinchDist.current = {
-					dist: Math.hypot(t1.clientX - t2.clientX, t1.clientY - t2.clientY),
-					cx: (t1.clientX + t2.clientX) / 2,
-					cy: (t1.clientY + t2.clientY) / 2,
-				};
-			}
-		},
-		[containerRef, activeYAxes, handleAutoScaleX, handleAutoScaleY],
-	);
+	const { handleWheel, handleMouseDown, handleTouchStart } = usePanZoomHandlers({
+		containerRef,
+		containerRectRef,
+		wheelTimeoutRef,
+		panStateRef,
+		smoothZoomRef,
+		zoomBoxStartRef,
+		shiftDownRef,
+		lastMousePos,
+		panTargetRef,
+		lastTouchTime,
+		lastTouchPos,
+		lastPinchDist,
+		width,
+		height,
+		padding,
+		activeYAxes,
+		setIsWheeling,
+		onPanEnd,
+		performZoom,
+		setIsZooming,
+		updateZoomBoxDom,
+		setPanTarget,
+		handleAutoScaleX,
+		handleAutoScaleY,
+	});
 
 	// Raw event listeners (non-React for passive:false touch)
 	useEffect(() => {
@@ -468,230 +394,14 @@ export function usePanZoom({
 		[],
 	);
 
-	// eslint-disable-next-line react-hooks/immutability
-	useEffect(() => {
-		let mouseMoveRaf = 0;
-		let pendingMouseEvent: {
-			clientX: number;
-			clientY: number;
-			shiftKey: boolean;
-		} | null = null;
-		const snapshotAxesToPanState = (ps: typeof panStateRef.current) => {
-			activeXAxes.forEach((a) => {
-				ps.startTargetX[a.id] = { ...targetXAxes.current[a.id] };
-			});
-			activeYAxes.forEach((a) => {
-				ps.startTargetY[a.id] = { ...targetYs.current[a.id] };
-			});
-		};
-		const handleSingleTouchPan = (e: TouchEvent, target: PanTarget) => {
-			if (e.cancelable) e.preventDefault();
-			const t = e.touches[0];
-			const ps = panStateRef.current;
-			if (!ps.active) {
-				ps.active = true;
-				if (lastTouchPos.current) {
-					ps.startX = lastTouchPos.current.x;
-					ps.startY = lastTouchPos.current.y;
-				}
-				ps.target = target;
-				snapshotAxesToPanState(ps);
-			}
-			ps.currentX = t.clientX;
-			ps.currentY = t.clientY;
-			updatePan();
-		};
-
-		const handlePinchZoom = (e: TouchEvent, target: PanTarget | null) => {
-			if (e.cancelable) e.preventDefault();
-			const rect =
-				containerRectRef.current ||
-				containerRef.current?.getBoundingClientRect();
-			if (!rect) return;
-			if (!lastPinchDist.current) return;
-			const gesture = computePinchGesture(
-				e.touches[0],
-				e.touches[1],
-				lastPinchDist.current.dist,
-			);
-			if (!gesture) return;
-			const { zfX, zfY, cx, cy, dist } = gesture;
-
-			// Apply pan
-			const panDx = cx - lastPinchDist.current.cx;
-			const panDy = cy - lastPinchDist.current.cy;
-
-			lastPinchDist.current = { dist, cx, cy };
-
-			performZoom(
-				{ x: zfX, y: zfY },
-				cx - rect.left,
-				cy - rect.top,
-				target || "all",
-				e.shiftKey,
-			);
-
-			// Apply pan AFTER performZoom overwrites the refs
-			if (
-				target === "all" ||
-				(target && typeof target === "object" && "xAxisId" in target)
-			) {
-				activeXAxes.forEach((a) => {
-					const cur = targetXAxes.current[a.id];
-					if (cur) {
-						targetXAxes.current[a.id] = panRangeByPixels(
-							cur.min,
-							cur.max,
-							-panDx,
-							chartWidth,
-						);
-					}
-				});
-			}
-
-			if (
-				target === "all" ||
-				(target && typeof target === "object" && "yAxisId" in target)
-			) {
-				activeYAxes.forEach((a) => {
-					const cur = targetYs.current[a.id];
-					if (cur) {
-						targetYs.current[a.id] = panRangeByPixels(
-							cur.min,
-							cur.max,
-							panDy,
-							chartHeight,
-						);
-					}
-				});
-			}
-
-			// Important: performZoom calls syncViewport() at the end.
-			// Since we modify the refs *after* performZoom, we must sync the viewport again
-			// to flush our pan adjustments to the actual component state.
-			syncViewport();
-		};
-
-		const handleTouchMoveRaw = (e: TouchEvent) => {
-			const target = panTargetRef.current;
-			if (e.touches.length === 1 && target && lastTouchPos.current) {
-				handleSingleTouchPan(e, target);
-			} else if (e.touches.length === 2 && lastPinchDist.current) {
-				handlePinchZoom(e, target);
-			}
-		};
-
-		const processMouseMove = () => {
-			mouseMoveRaf = 0;
-			const e = pendingMouseEvent;
-			pendingMouseEvent = null;
-			if (!e) return;
-			const rect =
-				containerRectRef.current ||
-				containerRef.current?.getBoundingClientRect();
-			if (!rect) return;
-			const mx = e.clientX - rect.left,
-				my = e.clientY - rect.top;
-
-			const target = panTargetRef.current;
-			// Only update hover state when not actively panning — saves work per frame.
-			if (!target && !zoomBoxStartRef.current) {
-				hoveredAxisIdRef.current = getHoveredYAxis(mx, my);
-				hoveredXAxisIdRef.current = getHoveredXAxis(mx, my);
-			}
-
-			if (zoomBoxStartRef.current) {
-				const box = zoomBoxStartRef.current;
-				box.endX = Math.max(padding.left, Math.min(width - padding.right, mx));
-				box.endY = Math.max(padding.top, Math.min(height - padding.bottom, my));
-				updateZoomBoxDom(box);
-				return;
-			}
-			if (!target || !lastMousePos.current) return;
-
-			shiftDownRef.current = e.shiftKey;
-
-			const ps = panStateRef.current;
-			if (!ps.active) {
-				ps.active = true;
-				ps.startX = lastMousePos.current.x;
-				ps.startY = lastMousePos.current.y;
-				ps.target = target;
-				snapshotAxesToPanState(ps);
-			}
-			ps.currentX = e.clientX;
-			ps.currentY = e.clientY;
-
-			updatePan();
-		};
-
-		const handleMouseMoveRaw = (e: MouseEvent) => {
-			pendingMouseEvent = {
-				clientX: e.clientX,
-				clientY: e.clientY,
-				shiftKey: e.shiftKey,
-			};
-			if (!mouseMoveRaf) {
-				mouseMoveRaf = requestAnimationFrame(processMouseMove);
-			}
-		};
-
-		const handleMouseUp = () => {
-			panStateRef.current.active = false;
-			containerRectRef.current = null;
-
-			if (zoomBoxStartRef.current) {
-				const box = zoomBoxStartRef.current;
-				zoomBoxStartRef.current = null;
-				setIsZooming(false);
-				const minX = Math.min(box.startX, box.endX);
-				const maxX = Math.max(box.startX, box.endX);
-				const minY = Math.min(box.startY, box.endY);
-				const maxY = Math.max(box.startY, box.endY);
-				if (maxX - minX > 5 && maxY - minY > 5) {
-					applyZoomBoxToAxes(
-						{ minX, maxX, minY, maxY },
-						activeXAxes,
-						activeYAxes,
-						width,
-						height,
-						padding,
-						targetXAxes.current,
-						targetYs.current,
-						isShiftPressedRef.current,
-					);
-					syncViewport();
-				}
-			}
-			onPanEnd();
-			setPanTarget(null);
-		};
-
-		const handleTouchEnd = () => {
-			onPanEnd();
-			setPanTarget(null);
-			lastTouchPos.current = null;
-			lastPinchDist.current = null;
-		};
-
-		window.addEventListener("mousemove", handleMouseMoveRaw);
-		window.addEventListener("mouseup", handleMouseUp);
-		window.addEventListener("touchmove", handleTouchMoveRaw, {
-			passive: false,
-		});
-		window.addEventListener("touchend", handleTouchEnd);
-		return () => {
-			window.removeEventListener("mousemove", handleMouseMoveRaw);
-			window.removeEventListener("mouseup", handleMouseUp);
-			window.removeEventListener("touchmove", handleTouchMoveRaw);
-			window.removeEventListener("touchend", handleTouchEnd);
-			if (mouseMoveRaf) cancelAnimationFrame(mouseMoveRaf);
-		};
-	}, [
+	usePanZoomEvents({
 		containerRef,
+		containerRectRef,
 		padding,
 		width,
 		height,
+		chartWidth,
+		chartHeight,
 		activeXAxes,
 		activeYAxes,
 		targetXAxes,
@@ -704,9 +414,18 @@ export function usePanZoom({
 		onPanEnd,
 		panStateRef,
 		updateZoomBoxDom,
-	chartWidth,
-			chartHeight,
-		]);
+		lastTouchPos,
+		lastPinchDist,
+		zoomBoxStartRef,
+		hoveredAxisIdRef,
+		hoveredXAxisIdRef,
+		shiftDownRef,
+		lastMousePos,
+		panTargetRef,
+		setPanTarget,
+		setIsZooming,
+		isShiftPressedRef,
+	});
 
 	usePanZoomKeyboard({
 		pressedKeys,
